+++ docs/DEVELOPMENT.md
@@ -43,7 +43,6 @@ export VITE_APP_SELF_SIGN_UP_ENABLED=<Self Signup Flag>
 export VITE_APP_MODEL_REGION=<Bedrock/SageMakerモデルのリージョン>
 export VITE_APP_MODEL_IDS=<Bedrock モデルの JSON Array>
 export VITE_APP_IMAGE_MODEL_IDS=<Bedrock 画像生成モデルの JSON Array>
-export VITE_APP_MODEL_FEATURE_FLAGS=<Bedrock モデルの Feature Flag JSON>
 export VITE_APP_ENDPOINT_NAMES=<SageMaker モデルの JSON Array>
 export VITE_APP_SAMLAUTH_ENABLED=<SAML 認証 Flag>
 export VITE_APP_SAML_COGNITO_DOMAIN_NAME=<SAML Cognito Domain>
@@ -68,7 +67,6 @@ export VITE_APP_SELF_SIGN_UP_ENABLED=true
 export VITE_APP_MODEL_REGION=us-west-2
 export VITE_APP_MODEL_IDS=["anthropic.claude-instant-v1","anthropic.claude-v2"]
 export VITE_APP_IMAGE_MODEL_IDS=["stability.stable-diffusion-xl-v1","amazon.titan-image-generator-v1"]
-export VITE_APP_MODEL_FEATURE_FLAGS=[...]
 export VITE_APP_ENDPOINT_NAMES=[]
 export VITE_APP_SAMLAUTH_ENABLED=true
 export VITE_APP_SAML_COGNITO_DOMAIN_NAME=your-preferred-name.auth.ap-northeast-1.amazoncognito.com
