+++ packages/cdk/lib/generative-ai-use-cases-stack.ts
@@ -158,7 +158,6 @@ export class GenerativeAiUseCasesStack extends Stack {
       webAclId: props.webAclId,
       modelRegion: api.modelRegion,
       modelIds: api.modelIds,
-      modelFeatureFlags: api.modelFeatureFlags,
       imageGenerationModelIds: api.imageGenerationModelIds,
       endpointNames: api.endpointNames,
       samlAuthEnabled,
@@ -279,10 +278,6 @@ export class GenerativeAiUseCasesStack extends Stack {
 
     new CfnOutput(this, 'ImageGenerateModelIds', {
       value: JSON.stringify(api.imageGenerationModelIds),
-    });
-
-    new CfnOutput(this, 'ModelFeatureFlags', {
-      value: JSON.stringify(api.modelFeatureFlags),
     });
 
     new CfnOutput(this, 'EndpointNames', {
