+++ packages/cdk/tsconfig.json
@@ -21,7 +21,7 @@
     "experimentalDecorators": true,
     "strictPropertyInitialization": false,
     "typeRoots": [
-      "./node_modules/@types"
+      "../../node_modules/@types"
     ],
     "types": [
       "node"
