+++ packages/common/package.json
@@ -1,0 +1,6 @@
+{
+    "name": "@generative-ai-use-cases-jp/common",
+    "version": "1.0.0",
+    "private": true,
+    "main": "src/index.ts"
+}
