+++ packages/web/src/vite-env.d.ts
@@ -17,7 +17,6 @@ interface ImportMetaEnv {
   readonly VITE_APP_MODEL_REGION: string;
   readonly VITE_APP_MODEL_IDS: string;
   readonly VITE_APP_IMAGE_MODEL_IDS: string;
-  readonly VITE_APP_MODEL_FEATURE_FLAGS: string;
   readonly VITE_APP_ENDPOINT_NAMES: string;
   readonly VITE_APP_SAMLAUTH_ENABLED: string;
   readonly VITE_APP_SAML_COGNITO_DOMAIN_NAME: string;
